+++ Src/Couchbase.Linq.UnitTests/Couchbase.Linq.UnitTests.csproj
@@ -120,6 +120,7 @@
     <Compile Include="QueryGeneration\ArrayIndexTests.cs" />
     <Compile Include="QueryGeneration\AggregateTests.cs" />
     <Compile Include="BucketQueryExecutorEmulator.cs" />
+    <Compile Include="QueryGeneration\ExpressionTransformers\StringComparisonExpressionTransformerTests.cs" />
     <Compile Include="QueryGeneration\NullHandlingTests.cs" />
     <Compile Include="QueryGeneration\N1QlHelpersTests.cs" />
     <Compile Include="QueryGeneration\NestTests.cs" />
