+++ Src/Couchbase.Linq.UnitTests/QueryGeneration/StringTests.cs
@@ -745,150 +745,5 @@ namespace Couchbase.Linq.UnitTests.QueryGeneration
         }
 
         #endregion
-
-        #region Compare Expression Conversion Tests
-
-        [Test]
-        public void StringCompare_LessThan1_ReturnsLessThanOrEqual()
-        {
-            var queryGenerationContext = new N1QlQueryGenerationContext()
-            {
-                MemberNameResolver =
-                    new JsonNetMemberNameResolver(new Newtonsoft.Json.Serialization.DefaultContractResolver()),
-                MethodCallTranslatorProvider = new DefaultMethodCallTranslatorProvider(),
-                Serializer = new Core.Serialization.DefaultSerializer()
-            };
-
-            var visitor = new Mock<N1QlExpressionTreeVisitor>(queryGenerationContext)
-            {
-                CallBase = true
-            };
-
-            var string1 = Expression.Constant("A");
-            var string2 = Expression.Constant("B");
-
-            BinaryExpression expression = Expression.LessThan(
-                Expression.Call(
-                    typeof(string).GetMethod("Compare", new[] { typeof(string), typeof(string) }),
-                    string1,
-                    string2
-                ),
-                Expression.Constant(1));
-
-            var newExpression = visitor.Object.ConvertStringCompareExpression(expression) as StringComparisonExpression;
-
-            Assert.NotNull(newExpression);
-            Assert.AreEqual(ExpressionType.LessThanOrEqual, newExpression.Operation);
-            Assert.AreEqual(string1, newExpression.Left);
-            Assert.AreEqual(string2, newExpression.Right);
-        }
-
-        [Test]
-        public void StringCompare_EqualTo1_ReturnsGreaterThan()
-        {
-            var queryGenerationContext = new N1QlQueryGenerationContext()
-            {
-                MemberNameResolver =
-                    new JsonNetMemberNameResolver(new Newtonsoft.Json.Serialization.DefaultContractResolver()),
-                MethodCallTranslatorProvider = new DefaultMethodCallTranslatorProvider(),
-                Serializer = new Core.Serialization.DefaultSerializer()
-            };
-
-            var visitor = new Mock<N1QlExpressionTreeVisitor>(queryGenerationContext)
-            {
-                CallBase = true
-            };
-
-            var string1 = Expression.Constant("A");
-            var string2 = Expression.Constant("B");
-
-            BinaryExpression expression = Expression.Equal(
-                Expression.Call(
-                    typeof(string).GetMethod("Compare", new[] { typeof(string), typeof(string) }),
-                    string1,
-                    string2
-                ),
-                Expression.Constant(1));
-
-            var newExpression = visitor.Object.ConvertStringCompareExpression(expression) as StringComparisonExpression;
-
-            Assert.NotNull(newExpression);
-            Assert.AreEqual(ExpressionType.GreaterThan, newExpression.Operation);
-            Assert.AreEqual(string1, newExpression.Left);
-            Assert.AreEqual(string2, newExpression.Right);
-        }
-
-        [Test]
-        public void StringCompare_GreaterThanNeg1_ReturnsGreaterThanOrEqual()
-        {
-            var queryGenerationContext = new N1QlQueryGenerationContext()
-            {
-                MemberNameResolver =
-                    new JsonNetMemberNameResolver(new Newtonsoft.Json.Serialization.DefaultContractResolver()),
-                MethodCallTranslatorProvider = new DefaultMethodCallTranslatorProvider(),
-                Serializer = new Core.Serialization.DefaultSerializer()
-            };
-
-            var visitor = new Mock<N1QlExpressionTreeVisitor>(queryGenerationContext)
-            {
-                CallBase = true
-            };
-
-            var string1 = Expression.Constant("A");
-            var string2 = Expression.Constant("B");
-
-            BinaryExpression expression = Expression.GreaterThan(
-                Expression.Call(
-                    typeof(string).GetMethod("Compare", new[] { typeof(string), typeof(string) }),
-                    string1,
-                    string2
-                ),
-                Expression.Constant(-1));
-
-            var newExpression = visitor.Object.ConvertStringCompareExpression(expression) as StringComparisonExpression;
-
-            Assert.NotNull(newExpression);
-            Assert.AreEqual(ExpressionType.GreaterThanOrEqual, newExpression.Operation);
-            Assert.AreEqual(string1, newExpression.Left);
-            Assert.AreEqual(string2, newExpression.Right);
-        }
-
-        [Test]
-        public void StringCompare_EqualToNeg1_ReturnsLessThan()
-        {
-            var queryGenerationContext = new N1QlQueryGenerationContext()
-            {
-                MemberNameResolver =
-                    new JsonNetMemberNameResolver(new Newtonsoft.Json.Serialization.DefaultContractResolver()),
-                MethodCallTranslatorProvider = new DefaultMethodCallTranslatorProvider(),
-                Serializer = new Core.Serialization.DefaultSerializer()
-            };
-
-            var visitor = new Mock<N1QlExpressionTreeVisitor>(queryGenerationContext)
-            {
-                CallBase = true
-            };
-
-            var string1 = Expression.Constant("A");
-            var string2 = Expression.Constant("B");
-
-            BinaryExpression expression = Expression.Equal(
-                Expression.Call(
-                    typeof(string).GetMethod("Compare", new[] { typeof(string), typeof(string) }),
-                    string1,
-                    string2
-                ),
-                Expression.Constant(-1));
-
-            var newExpression = visitor.Object.ConvertStringCompareExpression(expression) as StringComparisonExpression;
-
-            Assert.NotNull(newExpression);
-            Assert.AreEqual(ExpressionType.LessThan, newExpression.Operation);
-            Assert.AreEqual(string1, newExpression.Left);
-            Assert.AreEqual(string2, newExpression.Right);
-        }
-
-        #endregion
-
     }
 }
